refactor: move ArtifactoryRegistryTest and update references in RegistryTest
